+++ stream-chat-android-docs/src/main/kotlin/io/getstream/chat/docs/kotlin/client/cms/AndroidIntroduction.kt
@@ -35,7 +35,7 @@ class AndroidIntroduction {
 
         val statePluginFactory = StreamStatePluginFactory(
             config = StatePluginConfig(
-                backgroundSyncEnabled = true,
+                backgroundSyncEnabled = false,
                 userPresence = true,
             ),
             appContext = applicationContext