+++ stream-chat-android-state/src/main/java/io/getstream/chat/android/state/plugin/factory/StreamStatePluginFactory.kt
@@ -134,6 +134,7 @@ public class StreamStatePluginFactory(
             logicRegistry = logic,
             stateRegistry = stateRegistry,
             userPresence = config.userPresence,
+            isAutomaticSyncOnReconnectEnabled = config.isAutomaticSyncOnReconnectEnabled,
             syncMaxThreshold = config.syncMaxThreshold,
             now = { System.currentTimeMillis() },
         )