+++ stream-chat-android-ui-guides/src/main/java/io/getstream/chat/android/guides/GuidesApp.kt
@@ -32,7 +32,7 @@ class GuidesApp : Application() {
         val offlinePluginFactory = StreamOfflinePluginFactory(appContext = this)
         val statePluginFactory = StreamStatePluginFactory(
             config = StatePluginConfig(
-                backgroundSyncEnabled = true,
+                backgroundSyncEnabled = false,
                 userPresence = true,
             ),
             appContext = this,